+++ src/components/general/menus/MenuItem/index.ts
@@ -1,0 +1,1 @@
+export * from './MenuItem.component';
