+++ src/components/general/menus/PescaMenu/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaMenu.component';
