+++ src/components/general/menus/createPescaMenu.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import { PescaMenuContextType } from '@moneyboy/api/PescaMenuContextType';
-import { createPescaMenuItem } from '@moneyboy/components/general/menus/menuItem';
-import { createPescaMenuContainer } from '@moneyboy/components/general/menus/pescaMenu';
-import { createContext } from 'react';
+import { PescaMenuContextType } from "@moneyboy/api/PescaMenuContextType";
+import { createPescaMenuItem } from "@moneyboy/components/general/menus/MenuItem";
+import { createPescaMenuContainer } from "@moneyboy/components/general/menus/PescaMenu";
+import { createContext } from "react";
 
 /**
  * Create a PescaMenu.
